requirements.txt
python-binance>=1.0.16
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
ccxt>=3.0.0
pytest>=7.3.1
python-dotenv>=1.0.0
pydantic>=2.0.0
websocket-client>=1.5.1
aiohttp>=3.8.4
rich>=13.3.5
loguru>=0.7.0
schedule>=1.2.0
ta>=0.10.2
yfinance>=0.2.18
requests>=2.31.0
python-telegram-bot>=20.3
pytz>=2023.3 